retire wr_presentInviteActivityViewControllerWithSourceView
