+++ Wire-iOS/Sources/UserInterface/ContactsUI/ContactsViewController+Invite.swift
@@ -1,30 +1,33 @@
-// 
+
 // Wire
-// Copyright (C) 2016 Wire Swiss GmbH
-// 
+// Copyright (C) 2019 Wire Swiss GmbH
+//
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
-// 
+//
 // This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
-// 
+//
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see http://www.gnu.org/licenses/.
-// 
+//
 
+import Foundation
 
-#import <UIKit/UIKit.h>
+extension ContactsViewController {
+    @objc
+    func sendIndirectInvite(_ sender: UIView) {
+        let shareItemProvider = ShareItemProvider(placeholderItem: "")
+        let activityController = UIActivityViewController(activityItems: [shareItemProvider], applicationActivities: nil)
 
-#import "UIActivityViewController+Invite.h"
+        activityController.excludedActivityTypes = [UIActivity.ActivityType.airDrop]
 
+        activityController.configPopover(pointToView: sender)
 
-
-@interface UIViewController (WR_Invite)
-
-- (void)wr_presentInviteActivityViewControllerWithSourceView:(UIView *)sourceView logicalContext:(GenericInviteContext)logicalContext;
-
-@end
+        present(activityController, animated: true)
+    }
+}
