+++ Wire-iOS/Sources/UserInterface/Conversation/Content/ConversationContentViewController+MessageAction.swift
@@ -92,7 +92,7 @@ extension ConversationContentViewController {
                     targetView = view
                 }
 
-                if let saveController = UIActivityViewController(message: message, from: targetView) {
+                if let saveController = UIActivityViewController(message: message, from: targetView) {///TODO:
                     present(saveController, animated: true)
                 }
             }
