--- Wire-iOS/Sources/UserInterface/StartUI/Common/InviteActionSheet/UIActivityViewController+Invite.h
@@ -1,36 +1,0 @@
-// 
-// Wire
-// Copyright (C) 2016 Wire Swiss GmbH
-// 
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-// 
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-// 
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see http://www.gnu.org/licenses/.
-// 
-
-
-@import UIKit;
-
-typedef NS_ENUM(NSUInteger, GenericInviteContext) {
-    GenericInviteContextConversationList,
-    GenericInviteContextInvitesSearch,
-    GenericInviteContextStartUISearch,
-    GenericInviteContextStartUIBanner,
-    GenericInviteContextSettings
-};
-
-
-@interface UIActivityViewController (Invite)
-
-+ (instancetype)shareInviteActivityViewControllerWithCompletion:(UIActivityViewControllerCompletionWithItemsHandler)completion
-                                                 logicalContext:(GenericInviteContext)logicalContext;
-
-@end
